requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.0.8
djangorestframework==3.11.0
numpy==1.19.1
pandas==1.1.0
# psycopg2==2.7.3
psycopg2-binary
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1